[Mod] 클릭시 이벤트 발생하도록 수정
