--- src/js/login_function.js
@@ -1,22 +1,0 @@
-const signup = document.getElementById("sign-up");
-signin = document.getElementById("sign-in");
-loginin = document.getElementById("login-in");
-loginup = document.getElementById("login-up");
-
-signup.addEventListener("click", () => {
-    loginin.classList.remove("block");
-    loginup.classList.remove("none");
-
-    loginin.classList.add("none");
-    loginup.classList.add("block");
-})
-
-signin.addEventListener("click", () => {
-    loginin.classList.remove("none");
-    loginup.classList.remove("block");
-
-    loginin.classList.add("block");
-    loginup.classList.add("none");
-})
-
-// html파일과 연결 필요
